Replace Carton with cpanm to get acceptable pipline build times
While Carton is the low-maintenance solution it adds a lot of additional
run time to the CI build. The difference in time spent is because carton
builds a self-contained local/ directory. The replacement cpanm command
builds an incremental one, just downloading the ones which are not on
the system (or not new enough).