[gherkin-perl] Trim down on the size of the release tar ball
The 'dist' (Perl-speak for release archive) on 17.0.0 and 17.0.1 was
ca 8MB, due to the inclusion of test data, but also the perl5/
directory which contains the dependencies. Using MANIFEST.SKIP to
filter all these out of the release archive.

This commit trims down on the artifacts packaged: excluding all
artifacts that are author-only, such as acceptance/, helper-scripts/,
testdata/, etc.

Next to that, @aslak wanted the upload of the dist to be automatic,
not interrupted the question to proceed Y/N. This commit removes the
Dist::Zilla 'ConfirmRelease' plugin, which asks that question.

The minimum perl requirement declaration moved from dist.ini
(where it's ignored) to cpanfile (where it belongs).

CPAN wants a META.json file these days, next to the historically
required META.yml. Adding META.json for additonal compliance points.