[gherkin-perl] Regenerate CHANGELOG.md when the toplevel changed
Change the makefile to update CHANGELOG.md when the Gherkin one
updated. This fixes the bug in the 17.0.1 release where it wasn't
refreshed in the release.

Also, MetaCPAN can render Markdown these days, so there's no need
to store the markdown as CHANGES (which gets rendered as plain text).
And since MetaCPAN will look for CHANGELOG.md as well as CHANGES,
there's no reason to publish under a different name anymore.